refactor(ops_shared): lift find_elements orchestration into OpsShared
do_find_elements/4 was 35 lines of byte-identical orchestration in
CDPClient and BiDiClient — register query_id, encode ops as JSON,
fire bootstrap, await find result, branch lazy/eager — differing only
in 3 protocol-specific helpers (cast_register, fetch_element_refs,
final_sync_exec).

Move the orchestration + lazy_elements/3 + parent_object_id/1 helpers
into OpsShared as macro-injected find_elements/3 + find_elements_lazy/3
public API, plus a do_find_elements/5 module function that consumes
the client module atom to look up the 3 protocol-specific primitives.

Each client retains those 3 primitives as @doc-false public functions
(promoted from defp) so OpsShared can call them by module atom.

CDPClient -95, BiDiClient -74, OpsShared +119. Net -50, plus a single
source of truth for the find-flow.
